Add STOQS permalink to LRAUV deployment HTML index
- Refactor make_permalink: add stoqs_url_from_ds() and
  lrauv_stoqs_base_url() to generate a STOQS "Share this view" URL
  directly from an xarray Dataset; switch gen_permalink() from a
  parameter-id/clicks model to a platform_clicks selector; fix
  get_parameter_id() URL construction (Path() was mangling the
  query string)
- Refactor lrauv_deployment_plots: extract _build_and_write_html()
  from plot_deployment(); call stoqs_url_from_ds() and pass the
  resulting URL (or None on failure) to _write_html(); add a STOQS
  section with a "Share this view" link to the generated HTML
- Add OPeNDAP data-access-form (.nc.html) links for each log file
  in the HTML index
- Enable parallel/chunked open_mfdataset() with dask
- Fix np.linspace int cast in create_products.py
- Add dask and lzstring to pyproject.toml dependencies
- Update tests: fix test_platform_id_in_permalink (platform_clicks
  key, platform name); fix test_stoqs_failure_still_writes_html
  (patch stoqs_url_from_ds directly)

+++ src/data/create_products.py
@@ -578,7 +578,7 @@ class CreateProducts:
         idist = np.linspace(
             distnav.to_numpy()[0],
             distnav.to_numpy()[-1],
-            3 * self.ds["profile_number"].to_numpy()[-1],
+            int(3 * self.ds["profile_number"].to_numpy()[-1]),
         )
         # Vertical gridded to .5 m, rounded down to nearest 50m
         max_depth = np.floor(self.ds.cf["depth"].max() / 50) * 50
